# docs/contracts/public-rest.openapi.yaml
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/sessions/refresh": {
      "post": {
        "operationId": "Auth.refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "refreshToken"
                ],
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "writeOnly": true,
                    "maxLength": 512
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSession"
                }
              }
            }
          },
          "default": {
            "description": "Stable non-sensitive error envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Rotate a refresh session and issue fresh tokens",
        "tags": [
          "authentication"
        ],
        "x-swar-auth-kind": "PUBLIC",
        "x-swar-required-permissions": [],
        "x-swar-rate-limit-category": "SENSITIVE",
        "x-swar-idempotency": "NOT_APPLICABLE"
      }
    },
...
    "/api/v1/trusted-speakers": {
      "post": {
        "operationId": "TrustedSpeakers.create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTrustedSpeakerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrustedSpeaker"
                }
              }
            }
          },
          "default": {
            "description": "Stable non-sensitive error envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a tenant-scoped trusted-speaker record",
        "tags": [
          "trusted-speakers"
        ],
        "x-swar-auth-kind": "USER_ACCESS_JWT",
        "x-swar-required-permissions": [
          "enrollment.manage"
        ],
        "x-swar-rate-limit-category": "SENSITIVE",
        "x-swar-idempotency": "IDEMPOTENCY_KEY_REQUIRED"
      }
    },
    "/api/v1/trusted-speakers/{trustedSpeakerId}/consents": {
      "post": {
        "operationId": "TrustedSpeakers.consent",
        "parameters": [
          {
            "name": "trustedSpeakerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantConsentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Consent"
                }
              }
            }
          },
          "default": {
            "description": "Stable non-sensitive error envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Record explicit enrollment consent",
        "tags": [
          "trusted-speakers"
        ],
        "x-swar-auth-kind": "USER_ACCESS_JWT",
        "x-swar-required-permissions": [
          "enrollment.manage"
        ],
        "x-swar-rate-limit-category": "SENSITIVE",
        "x-swar-idempotency": "IDEMPOTENCY_KEY_REQUIRED"
      }
    },
    "/api/v1/enrollment-consents/{consentId}/revoke": {
      "post": {
        "operationId": "TrustedSpeakers.revoke",
        "parameters": [
          {
            "name": "consentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeConsentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResult"
                }
              }
            }
          },
          "default": {
            "description": "Stable non-sensitive error envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke consent and dependent active analysis",
        "tags": [
          "trusted-speakers"
        ],
        "x-swar-auth-kind": "USER_ACCESS_JWT",
        "x-swar-required-permissions": [
          "enrollment.manage"
        ],
        "x-swar-rate-limit-category": "SENSITIVE",
        "x-swar-idempotency": "IDEMPOTENCY_KEY_REQUIRED"
      }
    },
    "/api/v1/voiceprints/{voiceprintId}": {
      "delete": {
        "operationId": "TrustedSpeakers.deleteVoiceprint",
        "parameters": [
          {
            "name": "voiceprintId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResult"
                }
              }
            }
          },
          "default": {
            "description": "Stable non-sensitive error envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete encrypted voiceprint material and revoke dependent analysis",
        "tags": [
          "trusted-speakers"
        ],
        "x-swar-auth-kind": "USER_ACCESS_JWT",
        "x-swar-required-permissions": [
          "voiceprint.delete"
        ],
        "x-swar-rate-limit-category": "SENSITIVE",
        "x-swar-idempotency": "IDEMPOTENCY_KEY_REQUIRED"
      }
    },
    "/api/v1/voice-enrollments": {
      "post": {
        "operationId": "VoiceEnrollment.enroll",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "trustedSpeakerId",
                  "consentId",
                  "expectedModelVersionId",
                  "declaredDurationsMs",
                  "samples"
                ],
                "properties": {
                  "trustedSpeakerId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "consentId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "expectedModelVersionId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "declaredDurationsMs": {
                    "type": "string"
                  },
                  "samples": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Voiceprint"
                }
              }
            }
          },
          "default": {
            "description": "Stable non-sensitive error envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create an encrypted voiceprint from transient consented samples",
        "tags": [
          "voice-enrollment"
        ],
        "x-swar-auth-kind": "USER_ACCESS_JWT",
        "x-swar-required-permissions": [
          "enrollment.manage"
        ],
        "x-swar-rate-limit-category": "SENSITIVE",
        "x-swar-idempotency": "IDEMPOTENCY_KEY_REQUIRED"
      }
    },
...
    "/api/v1/interventions/{interventionId}/verification-challenges": {
      "post": {
        "operationId": "Interventions.request",
        "parameters": [
          {
            "name": "interventionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVerificationChallengeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationChallenge"
                }
              }
            }
          },
          "default": {
            "description": "Stable non-sensitive error envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Request independent step-up or callback verification",
        "tags": [
          "interventions"
        ],
        "x-swar-auth-kind": "USER_ACCESS_JWT",
        "x-swar-required-permissions": [
          "intervention.resolve"
        ],
        "x-swar-rate-limit-category": "SENSITIVE",
        "x-swar-idempotency": "IDEMPOTENCY_KEY_REQUIRED"
      }
    },
...
  },
  "info": {
    "title": "SWAR REST and callback API",
    "description": "Versioned contracts for authenticated users, internal services, and signed LiveKit webhooks. Examples are fictional.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1",
      "description": "Versioned SWAR API base path"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "mlService": {
        "type": "http",
        "scheme": "bearer"
      },
      "verifierService": {
        "type": "http",
        "scheme": "bearer"
      },
      "liveKitWebhook": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    },
    "schemas": {
...
      "CreateTrustedSpeakerDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "maxLength": 160
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "externalReference": {
            "type": "string",
            "maxLength": 160
          }
        },
        "required": [
          "label"
        ]
      },
      "GrantConsentDto": {
        "type": "object",
        "properties": {
          "purposeCode": {
            "type": "string",
            "example": "expected-speaker-verification"
          },
          "noticeVersion": {
            "type": "string",
            "example": "notice-v1"
          },
          "consentAffirmed": {
            "type": "boolean",
            "enum": [
              true
            ],
            "description": "Must be explicitly true."
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "purposeCode",
          "noticeVersion",
          "consentAffirmed"
        ]
      },
      "RevokeConsentDto": {
        "type": "object",
        "properties": {
          "reasonCode": {
            "type": "string",
            "maxLength": 80
          }
        },
        "required": [
          "reasonCode"
        ]
      },
...
      "CreateVerificationChallengeDto": {
        "type": "object",
        "properties": {
          "method": {
            "type": "string",
            "enum": [
              "CALLBACK",
              "OUT_OF_BAND"
            ]
          }
        },
        "required": [
          "method"
        ]
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "AuthSession": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "accessToken",
          "refreshToken",
          "tokenType",
          "expiresIn",
          "principal"
        ],
        "properties": {
          "accessToken": {
            "type": "string",
            "readOnly": true,
            "x-swar-sensitive": true
          },
          "refreshToken": {
            "type": "string",
            "readOnly": true,
            "x-swar-sensitive": true
          },
          "tokenType": {
            "type": "string",
            "enum": [
              "Bearer"
            ]
          },
          "expiresIn": {
            "type": "integer",
            "minimum": 1
          },
          "principal": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "userId",
              "membershipId",
              "organizationId",
              "deviceId",
              "roles"
            ],
            "properties": {
              "userId": {
                "type": "string",
                "format": "uuid"
              },
              "membershipId": {
                "type": "string",
                "format": "uuid"
              },
              "organizationId": {
                "type": "string",
                "format": "uuid"
              },
              "deviceId": {
                "type": "string",
                "format": "uuid"
              },
              "roles": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "TrustedSpeaker": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "trustedSpeakerId",
          "status",
          "label"
        ],
        "properties": {
          "trustedSpeakerId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "label": {
            "type": "string"
          }
        }
      },
      "Consent": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "consentId",
          "trustedSpeakerId",
          "status",
          "expiresAt"
        ],
        "properties": {
          "consentId": {
            "type": "string",
            "format": "uuid"
          },
          "trustedSpeakerId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "StatusResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "consentId": {
            "type": "string",
            "format": "uuid"
          },
          "voiceprintId": {
            "type": "string",
            "format": "uuid"
          },
          "interventionId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
      "Voiceprint": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "voiceprintId",
          "trustedSpeakerId",
          "modelVersionId",
          "status"
        ],
        "properties": {
          "voiceprintId": {
            "type": "string",
            "format": "uuid"
          },
          "trustedSpeakerId": {
            "type": "string",
            "format": "uuid"
          },
          "modelVersionId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "VerificationChallenge": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "challengeId",
          "interventionId",
          "status",
          "method",
          "expiresAt"
        ],
        "properties": {
          "challengeId": {
            "type": "string",
            "format": "uuid"
          },
          "interventionId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "enum": [
              "PENDING",
              "PASSED",
              "FAILED",
              "EXPIRED",
              "CANCELLED"
            ]
          },
          "method": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "resultCode": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}